rename : hooks 내부 구조화
- 디렉토리 구조 정리
- shared 이상 계층을 참조 하는 모듈은 분리 - 수정예정
