+++ client/src/shared/hooks/index.ts
@@ -1,14 +1,14 @@
 export { default as useAuth } from './useAuth';
-export { default as useCheckList } from './useCheckList';
-export { default as useDebounce } from './useDebounce';
-export { default as useMediaRecord } from './useMediaRecord';
+export { default as useCheckList } from './checklist/useCheckList';
+export { default as useDebounce } from './debounce/useDebounce';
+export { default as useMediaRecord } from './record/useMediaRecord';
 export { default as useModal } from './useModal';
 export { default as usePageNation } from './usePageNation';
-export { default as usePwCheck } from './usePwCheck';
-export { default as useSectionObserver } from './useSectionObserver';
+export { default as usePwCheck } from './pwcheck/usePwCheck';
+export { default as useSectionObserver } from './sectionobserver/useSectionObserver';
 export { default as useThemeToggler } from './useThemeToggler';
 export { default as useToast } from './useToast';
-export { default as useWebSpeech } from './useWebSpeech';
+export { default as useWebSpeech } from './speech/useWebSpeech';
 export { default as useDragCorrection } from './useDragCorrection';
-export { default as useInterval } from './useInterval';
-export { default as useCounter } from './useCounter';
+export { default as useInterval } from './interval/useInterval';
+export { default as useCounter } from './counter/useCounter';
