+++ client/src/shared/hooks/sectionobserver/useSectionObserver.ts
