+++ client/src/shared/hooks/speech/useWebSpeech.ts
